==> src/6_gera_embeddings_js.ipynb <==
## 1. Carrega as bases de dados

#!pip install --trusted-host pypi.org --trusted-host files.pythonhosted.org transformers
#!pip install --trusted-host pypi.org --trusted-host files.pythonhosted.org torch
#!pip install --trusted-host pypi.org --trusted-host files.pythonhosted.org tensorflow

import pandas as pd

###############################################################################
# Modelos
#MODELO = 'rufimelo/Legal-BERTimbau-sts-large-ma-v3'
#MAX_SEQ_LENGTH = 512

#MODELO = 'sentence-transformers/paraphrase-multilingual-mpnet-base-v2'
#MAX_SEQ_LENGTH = 128

#MODELO = 'neuralmind/bert-large-portuguese-cased'
#MAX_SEQ_LENGTH = 512

#MODELO = 'stjiris/bert-large-portuguese-cased-legal-mlm-sts-v1.0'
#MAX_SEQ_LENGTH = 512

#MODELO = 'stjiris/bert-large-portuguese-cased-legal-mlm-nli-sts-v1'
#MAX_SEQ_LENGTH = 512

#MODELO = 'Luciano/bert-base-portuguese-cased-finetuned-tcu-acordaos'
#MAX_SEQ_LENGTH = 512

MODELO = 'neuralmind/bert-base-portuguese-cased'
MAX_SEQ_LENGTH = 512

###############################################################################

CAMINHO_MODELO = MODELO.split("/")[-1]

# Pasta com os dados
PASTA_DADOS = './dados/'

# Pasta com os dados de jurisprudência já tratados
PASTA_JURIS_TCU = f'{PASTA_DADOS}outputs/1_tratamento_juris_tcu/'

# Pasta onde serão armazenados os resultados desse caderno
PASTA_RESULTADO_CADERNO = f'{PASTA_DADOS}outputs/6_gera_embeddings_js/{CAMINHO_MODELO}/'

# Tamanho do lote
#QUANTIDADE_DE_ITENS_CARREGADOS = 5

# Tamanho do lote
SOBRESCREVER_EMBEDDINGS = False

# Tamanho do lote
TAMANHO_DO_LOTE = 512

# Função que carrega os arquivos 
def carrega_juris_tcu():
    doc1 = pd.read_csv(f'{PASTA_JURIS_TCU}doc_tratado_parte_1.csv', sep='|')
    doc2 = pd.read_csv(f'{PASTA_JURIS_TCU}doc_tratado_parte_2.csv', sep='|')
    doc3 = pd.read_csv(f'{PASTA_JURIS_TCU}doc_tratado_parte_3.csv', sep='|')
    doc4 = pd.read_csv(f'{PASTA_JURIS_TCU}doc_tratado_parte_4.csv', sep='|')
    doc = pd.concat([doc1, doc2, doc3, doc4], ignore_index=True)

    return doc

from formatador import remove_html

# Carrega os arquivos para doc
doc = carrega_juris_tcu()
#doc = doc.head(QUANTIDADE_DE_ITENS_CARREGADOS)

#Remove tags do enunciado
doc['ENUNCIADO'] = doc['ENUNCIADO'].apply(remove_html)

#Transforma dataframe em dicionário
doc = doc.to_dict(orient='list')
print(doc.keys())

## 2. Tokenização do texto.

from transformers import AutoTokenizer

#Carregando o nokenizador Legal BERTimbau V3
model_ckpt = MODELO

tokenizer = AutoTokenizer.from_pretrained(model_ckpt)

#Definição da função que realizará a tokenização em lotes
def tokenize(batch):
    return tokenizer(batch["ENUNCIADO"], padding=True, truncation=True, return_tensors='pt', max_length=MAX_SEQ_LENGTH)

# Aplicando a função de tokenização aos dados
doc_encoded = doc.copy()
tokenized_outputs = tokenize(doc_encoded)
tokenized_outputs

tokenized_outputs.input_ids.shape

# Armazenando input_ids, attention_mask e token_type_ids em doc_encoded
doc_encoded['input_ids'] = tokenized_outputs['input_ids']
doc_encoded['attention_mask'] = tokenized_outputs['attention_mask']
doc_encoded['token_type_ids'] = tokenized_outputs['token_type_ids']
doc_encoded.keys()

# Verifica se a tokenização foi realizada adequadamente
from itertools import islice

for enunciado, input_id in islice(zip(doc_encoded["ENUNCIADO"], doc_encoded["input_ids"]), 3):
    print(f"Enunciado: {enunciado}, \nInput IDs: {tokenizer.convert_ids_to_tokens(input_id)}\n")

## 3. Obtenção dos embeddings.

import torch
from transformers import AutoModel

# Carrega modelo

#Caso exista GPU utilize-a, caso contrário use a CPU
device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
model = AutoModel.from_pretrained(model_ckpt).to(device)

# Função para agregaçção da última camada oculta pela média
def mean_pooling(model_output, attention_mask):
    token_embeddings = model_output[0] #First element of model_output contains all token embeddings
    input_mask_expanded = attention_mask.unsqueeze(-1).expand(token_embeddings.size()).float()
    return torch.sum(token_embeddings * input_mask_expanded, 1) / torch.clamp(input_mask_expanded.sum(1), min=1e-9)

# Função para extração da última camada oculta
def extract_hidden_states(batch):
    # Place model inputs on the GPU
    inputs = {k:v.to(device) for k,v in batch.items() 
              if k in tokenizer.model_input_names}
    # Extract last hidden states
    with torch.no_grad():
        model_output = model(**inputs)
    
    return batch['KEY'], model_output.last_hidden_state[:,0].cpu().numpy(), mean_pooling(model_output, inputs['attention_mask']).cpu().numpy()

# Função para dividir dicionário em lotes
def dividir_dicionario_em_lotes(dicionario, tamanho_do_lote):
    vetor_de_dicionarios = []
    max_len = max(len(v) for v in dicionario.values())  # Encontrando o vetor de valores mais longo
    
    for i in range(0, max_len, tamanho_do_lote):
        novo_dicionario = {chave: valores[i:i + tamanho_do_lote] for chave, valores in dicionario.items()}
        vetor_de_dicionarios.append(novo_dicionario)
    
    return vetor_de_dicionarios

import numpy as np

# Função para reconstruir dicionário a partir de lotes
def reconstruir_dicionario_a_partir_de_lotes(vetor_de_dicionarios):
    dicionario_reconstruido = {}
    
    # Inicializando listas vazias para cada chave no primeiro dicionário do vetor
    for chave in vetor_de_dicionarios[0].keys():
        dicionario_reconstruido[chave] = []
    
    # Iterando sobre cada dicionário no vetor e concatenando os valores para cada chave
    for dicionario_lote in vetor_de_dicionarios:
        for chave, valores in dicionario_lote.items():
            dicionario_reconstruido[chave].extend(valores)
    
    # Transforma numpy array em tensor
    dicionario_reconstruido['cls_hidden_state'] = torch.tensor(np.array(dicionario_reconstruido['cls_hidden_state']))
    dicionario_reconstruido['mean_hidden_state'] = torch.tensor(np.array(dicionario_reconstruido['mean_hidden_state']))
    
    return dicionario_reconstruido

CAMINHO_MODELO

# Processar e salvar embeddings
import pickle
from tqdm import tqdm
import os

# Divide doc_encoded em lotes
doc_encoded_em_lotes = dividir_dicionario_em_lotes(doc_encoded, TAMANHO_DO_LOTE)

# Processa e salva embeddings
for i, dicionario in enumerate(tqdm(doc_encoded_em_lotes), start=1):
    
    caminho_arquivo = f'{PASTA_RESULTADO_CADERNO}{CAMINHO_MODELO}_embeddings_js_{i}.pickle'
    if  not SOBRESCREVER_EMBEDDINGS and os.path.exists(caminho_arquivo):
        continue

    key, cls_hidden_state, mean_hidden_state = extract_hidden_states(dicionario)
    
    # Cria estrutura que será salva em arquivo
    embeddings_js = {
        'key': key,
        'cls_hidden_state': cls_hidden_state,
        'mean_hidden_state': mean_hidden_state
    }
    
    # Gravando lote em um arquivo .pickle
    with open(caminho_arquivo, 'wb') as arquivo_pickle:
        pickle.dump(embeddings_js, arquivo_pickle)

dicionario['input_ids'].size()
#dicionario['attention_mask']
#dicionario['token_type_ids']

doc_encoded_em_lotes[0]['KEY']

## 4. Cálculo da distância entre os embeddings.

# Função para restaurar embeddings dos arquivos pickle

import os
import pickle

def restaurar_doc_encoded_de_pickle(pasta_resultado_caderno):
    # Lista para armazenar os dicionários lidos dos arquivos .pickle
    doc_encoded_restaurado = []

    # Listando todos os arquivos .pickle no diretório especificado
    arquivos_pickle = [arq for arq in os.listdir(pasta_resultado_caderno) if arq.endswith('.pickle')]

    # Ordenando os arquivos pelo número (assumindo que os nomes dos arquivos seguem o padrão embeddings_js_X.pickle)
    arquivos_pickle.sort(key=lambda x: int(x.split('_')[-1].split('.')[0]))

    # Lendo cada arquivo .pickle e restaurando o dicionário
    for nome_arquivo in arquivos_pickle:
        caminho_arquivo = os.path.join(pasta_resultado_caderno, nome_arquivo)
        with open(caminho_arquivo, 'rb') as arquivo_pickle:
            dicionario_restaurado = pickle.load(arquivo_pickle)
            doc_encoded_restaurado.append(dicionario_restaurado)

    return reconstruir_dicionario_a_partir_de_lotes(doc_encoded_restaurado)

doc_encoded_restaurado = restaurar_doc_encoded_de_pickle(PASTA_RESULTADO_CADERNO)

import numpy as np

doc_hidden = doc_encoded.copy()
doc_hidden['cls_hidden_state'] = doc_encoded_restaurado['cls_hidden_state']
doc_hidden['mean_hidden_state'] = doc_encoded_restaurado['mean_hidden_state']
print(f"cls_hidden_state: {doc_hidden['cls_hidden_state'].size()}\n")
print(f"mean_hidden_state: {doc_hidden['mean_hidden_state'].size()}")

# Extraindo os embeddings de duas frases
embedding1_tensor = doc_hidden['mean_hidden_state'][0]
embedding2_tensor = doc_hidden['mean_hidden_state'][4]

# Normalizando os embeddings
embedding1_norm = embedding1_tensor / embedding1_tensor.norm()
embedding2_norm = embedding2_tensor / embedding2_tensor.norm()

# Calculando a similaridade por cosseno
cosine_similarity = torch.dot(embedding1_norm, embedding2_norm)

print(f"Similaridade por cosseno: {cosine_similarity.item()}")

doc_encoded_restaurado

print(doc_hidden['ENUNCIADO'][0])
print(doc_hidden['ENUNCIADO'][1])
print(doc_hidden['ENUNCIADO'][4])
